refactor: change typescript style to camel
