+++ .gitignore
@@ -6,7 +6,7 @@
 
 */**/node_modules
 src/**/*.js
-src/local_config.ts
+src/localConfig.ts
 !src/open_digger.js
 src/cron/tasks/local_*
 */**/.ipynb_checkpoints
